requirements.txt
flask
mysql-connector-python
opencv-python
pytesseract
python-dotenv
pandas
fpdf
